API/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.6.1
scikit-learn==1.5.1
joblib==1.4.2
numpy==1.26.4
python-multipart==0.0.9
google-generativeai==0.7.2
